+++ build.gradle
@@ -56,7 +56,7 @@ dependencies {
 	implementation 'org.ow2.asm:asm-util:7.1'
 
 	// Patchwork
-	implementation 'com.patchworkmc:patchwork-manifest:0.1.1'
+	implementation 'com.patchworkmc:patchwork-manifest:0.1.2'
 	implementation 'com.patchworkmc:application-core:0.1.0'
 
 	// Fukkit Gloom (for access transformers)
