+++ src/main/java/com/patchworkmc/jar/ForgeModJar.java
@@ -1,10 +1,12 @@
 package com.patchworkmc.jar;
 
 import java.nio.file.Path;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import com.patchworkmc.manifest.accesstransformer.AccessTransformerList;
-import com.patchworkmc.manifest.converter.com.patchworkmc.manifest.AccessTransformerListMerger;
+import com.patchworkmc.manifest.AccessTransformerListMerger;
 import com.patchworkmc.manifest.mod.ModManifest;
 import com.patchworkmc.manifest.mod.ModManifestEntry;
 
@@ -12,6 +14,7 @@ public class ForgeModJar {
 	private Path jarPath;
 	private ModManifest manifest;
 	private AccessTransformerList accessTransformers;
+	private Set<String> dependencies = new HashSet<>();
 
 	public ForgeModJar(Path jarPath, ModManifest manifest, AccessTransformerList list) {
 		this.jarPath = jarPath;
@@ -20,22 +23,22 @@ public class ForgeModJar {
 	}
 
 	public void addDependencyJars(List<ForgeModJar> modJars) {
-		for (ForgeModJar mod : modJars) {
-			if (mod.equals(this)) {
+		for (ForgeModJar proposedDependencyJar : modJars) {
+			if (proposedDependencyJar.equals(this)) {
 				continue;
 			}
 
 			boolean depends = false;
 
-			for (ModManifestEntry modManifestEntry : getManifest().getMods()) {
-				if (manifest.getDependencyMapping().containsKey(modManifestEntry.getModId())) {
+			for (ModManifestEntry modManifestEntry : proposedDependencyJar.getManifest().getMods()) {
+				if (this.manifest.getDependencyMapping().containsKey(modManifestEntry.getModId())) {
 					depends = true;
 					break;
 				}
 			}
 
 			if (depends) {
-				this.accessTransformers = AccessTransformerListMerger.createMergedList(this.accessTransformers, mod.getAccessTransformers());
+				this.accessTransformers = AccessTransformerListMerger.createMergedList(this.accessTransformers, proposedDependencyJar.getAccessTransformers());
 			}
 		}
 	}
